requirements.txt
openai
python-dotenv
streamlit
Pillow
pymupdf
sendgrid
pytesseract
requests